mri-analyzer-service/requirements.txt
fastapi
uvicorn
cognee
python-multipart
python-dotenv
requests
fpdf
